Completed provider login - services page

+++ server/controllers/locationController.js
@@ -1,5 +1,8 @@
 const express = require('express');
 const { Location, State } = require('../models/masterModels');
+const { ServiceRequest } = require('../models/serviceModels');
+const { Provider, ProviderService } = require('../models/providerModel'); // Import the provider model
+
 
 
 // to get all Locations (for admin purpose)
@@ -12,7 +15,14 @@ const getAllLocations = async (req, res, next) => {
     // Slice the Locations array to get the paginated data
 
     try {
-        const Locations = await Location.find().limit(limit).skip(startIndex);
+        const Locations = await Location.find().limit(limit).skip(startIndex).populate('stateId', 'name');;
+
+        // // Add custom field merging state name and location name
+        // const locationsWithMergedNames = Locations.map(location => ({
+        //     ...location._doc,
+        //     mergedName: `${location.stateId.name} - ${location.name}`
+        // }));
+        // console.log(locationsWithMergedNames);
         res.status(200).json({
             data: Locations,
             page: parseInt(page),
@@ -103,16 +113,26 @@ const getLocationById = async (req, res, next) => {
 //to Create a new Location
 const createLocation = async (req, res, next) => {
     const { name, code, pincode, isActive, stateId } = req.body;
-    if (!name || !stateId || !code) {
-        return res.status(400).json({ message: "Name, State and Code is required" });
-    }
-    if (Location.find(Location => Location.stateId === stateId && (Location.name === name || Location.code === code))) {
-        return res.status(400).json({ message: "Name or Code is exists." });
+    // console.log(req.body);
+    if (!name || !pincode || !stateId) {       //|| !code
+        return res.status(422).json({ message: "Name, State and Pincode is required", success: false });
     }
     try {
 
-        const newLocation = new Location({ name: name, code: code, pincode: pincode, isActive: isActive, stateId: stateId });
+        const existingLocation = await Location.findOne({ name: name, stateId: stateId });
+        if (existingLocation) {
+            return res.status(406).json({ message: "Name or Code already exists.", success: false });
+        }
+
+        const newLocation = new Location({
+            name: name, pincode: pincode, isActive: isActive, stateId: stateId,
+            createdBy: req.user.id,
+            UpdatedBy: req.user.id
+        }); //code: code,
         await newLocation.save();
+
+        await newLocation.populate('stateId', 'name')
+
         return res.status(200).json({
             message: "New Location created Successfully",
             data: newLocation,
@@ -129,21 +149,39 @@ const createLocation = async (req, res, next) => {
 const updateLocation = async (req, res, next) => {
     const _id = req.params.id;
     const { name, code, pincode, isActive, stateId } = req.body;
+    // console.log(req.body);
     if (!name || !stateId || !code) {
         return res.status(400).json({ message: "Name, State and Code is required" });
     }
-    if (Location.find(Location => Location.stateId === stateId && (Location.name === name || Location.code === code))) {
-        return res.status(400).json({ message: "Name or Code is exists." });
-    }
-    try {
-        const Location = Location.findByIdAndUpdate({ id: _id }, { name: name, code: code, pincode: pincode, isActive: isActive, stateId: stateId });
-        if (!Location) {
+
+    try { 
+         // const existingLocation = await Location.findOne({ $and: [{ name: name, stateId: stateId }, { _id: { $ne: _id } }] });
+        const existingLocation = await Location.findOne({
+            $and: [
+              { name: { $regex: `^${name.trim()}$`, $options: 'i' } },
+              { stateId: stateId },
+              { _id: { $ne: _id } }
+            ]
+          });
+       
+        if (existingLocation) {
+            return res.status(406).json({ message: "Location Name or Code already exists.", success: false });
+        }
+
+        const updatedLocation = await Location.findByIdAndUpdate(_id  , { name: name, code: code, pincode: pincode, isActive: isActive, stateId: stateId });
+        if (!updatedLocation) {
             return res.status(404).json({ message: `Location with id : ${_id} not found.` });
         }
-        await Location.save();
+
+        console.log("updateLocation: ", updatedLocation.validateSync());
+
+        // await updateLocation.save();
+
+        await updatedLocation.populate('stateId', 'name');
+
         return res.status(201).json({
             message: "Location updated Successfully",
-            data: Location,
+            data: updatedLocation,
             success: true
         });
 
@@ -153,15 +191,60 @@ const updateLocation = async (req, res, next) => {
 
 }
 
+// to update location status (active/inactive)
+const updateLocationStatus = async (req, res, next) => {
+    const _id = req.params.id;
+    const { isActive } = req.body;
+
+    if (isActive === undefined || isActive === null) {
+        return res.status(400).json({ message: "isActive status is required" });
+    }
+    const existingLocation = await Location.findById(_id);
+    if (!existingLocation) {
+        return res.status(406).json({ message: "Location is not exists.", success: false });
+    }
+    try {
+        const location = await Location.findByIdAndUpdate({ _id: _id }, { isActive: isActive });
+        if (!location) {
+            return res.status(404).json({ message: `Location with id : ${_id} not found.` });
+        }
+        // await Location.save();
+        return res.status(201).json({
+            message: "Location updated Successfully",
+            data: location,
+            success: true
+        });
+
+    } catch (error) {
+        console.log("Error in updateLocationStatus: ", error);
+        next(error)
+    }
+}
+
+
 //to delete a Location
 const deleteLocation = async (req, res, next) => {
     const _id = req.params.id;
     try {
-        const delLocation = Location.findByIdAndDelete(_id);
+        //check any records for this location in other collections
+        const serviceRequest = await ServiceRequest.findOne({ locationId: _id });
+        if (serviceRequest) {
+            return res.status(406).json({ message: "Location is already assigned to Service Request.", success: false });
+        }
+        const providers = await Provider.findOne({ locationId: _id });
+        if (providers) {
+            return res.status(406).json({ message: "Location is already assigned to Provider.", success: false });
+        }
+        const providerService = await ProviderService.findOne({ locationId: _id });
+        if (providerService) {
+            return res.status(406).json({ message: "Location is already assigned to Provider Service.", success: false });
+        }
+
+        const delLocation = await Location.findByIdAndDelete(_id);
         if (!deleteLocation) {
             return res.status(404).json({ message: `Location with id : ${_id} not found.` });
         }
-        await delLocation.save();
+        // await delLocation.save();
         return res.status(201).json({
             message: "Location deleted Successfully",
             data: delLocation,
@@ -180,5 +263,6 @@ module.exports = {
     getAlllocationsWithState,
     createLocation,
     updateLocation,
-    deleteLocation
+    deleteLocation,
+    updateLocationStatus
 }
