+++ server/controllers/stateController.js
@@ -1,10 +1,12 @@
 const express = require('express');
 // const { model } = require('mongoose');
-const { State } = require('../models/masterModels');
+const mongoose = require('mongoose'); // Add this line
+const { State, Location } = require('../models/masterModels');
+// const masterModels = require('../models/masterModels');
 
 
 // to get all states (for admin purpose)
-const getAllStates =  async (req, res, next)  => {
+const getAllStates = async (req, res, next) => {
 
     const { page = 1, limit = 10 } = req.query; // Default values: page 1, limit 10
     // find the starting and ending indices
@@ -22,7 +24,7 @@ const getAllStates =  async (req, res, next)  => {
             totalPages: Math.ceil(states.length / limit),
             message: "States retrieved Successfully"
         });
-        
+
     } catch (error) {
         next(error);
         // res.status(400).json(error);
@@ -31,7 +33,7 @@ const getAllStates =  async (req, res, next)  => {
 
 
 // to get only active states 
-const getAllActiveStates =  async (req, res, next) =>{
+const getAllActiveStates = async (req, res, next) => {
 
     const { page = 1, limit = 10 } = req.query; // Default values: page 1, limit 10
     // find the starting and ending indices
@@ -40,7 +42,7 @@ const getAllActiveStates =  async (req, res, next) =>{
     // Slice the states array to get the paginated data
 
     try {
-        const states = await State.find({isActive: true}).limit(limit).skip(startIndex);
+        const states = await State.find({ isActive: true }).limit(limit).skip(startIndex);
         res.status(200).json({
             data: states,
             page: parseInt(page),
@@ -49,7 +51,7 @@ const getAllActiveStates =  async (req, res, next) =>{
             totalPages: Math.ceil(states.length / limit),
             message: "States retrieved Successfully"
         });
-        
+
     } catch (error) {
         next(error);
         // res.status(400).json(error);
@@ -58,7 +60,7 @@ const getAllActiveStates =  async (req, res, next) =>{
 
 
 //to get single State details
-const getStateById =   async (req, res, next) =>{ 
+const getStateById = async (req, res, next) => {
     const _id = req.params.id;
     const state = State.find(state => state._id === _id);
     if (!state) {
@@ -69,18 +71,42 @@ const getStateById =   async (req, res, next) =>{
 
 
 //to Create a new State
-const createState =  async (req, res, next) =>{
-    const {name,  code, description, isActive } = req.body;
-    if( !name ||  !code ){
-        return res.status(400).json( { message: "Name and Code is required"});
-    }
-    if( State.find( state => state.name === name || state.code === code  ) ){
-        return res.status(400).json( { message: "Name or Code is exists."});
-    }
+const createState = async (req, res, next) => {
+    const { name, description, isActive } = req.body;
+    let code = req.body.code || "";
+    // console.log(req.body); 
     try {
+
+        if (!name || !description) {
+            return res.status(422).json({ message: "Name and description is required", success: false });
+        }
+
+        const existingState = await State.findOne({ $or: [{ name: name }, { code: code }] });
+        if (existingState) {
+            return res.status(406).json({ message: "Name or Code already exists.", success: false });
+        }
+   
+
+        const newState = new State({
+            name,
+            code,
+            description,
+            isActive,
+            createdBy: req.user.id,
+            updatedBy: req.user.id
+        }); 
         
-        const newState = new State ({ name: name, code: code, description: description , isActive: isActive }); 
+        // console.log("newState: ", newState); 
+        // mongoose.set('debug', true);
+        // console.log(newState.validateSync());
+        // console.log("newState: ", newState.validateSync());
+
         await newState.save();
+ 
+        if (!newState) {
+            return res.status(406).json({ message: "newState error", success: false });
+        }
+ 
         return res.status(200).json({
             message: "New State created Successfully",
             data: newState,
@@ -88,30 +114,39 @@ const createState =  async (req, res, next) =>{
         });
 
     } catch (error) {
+        console.log(error);
         next(error)
     }
-   
+
 }
 
 //to Update a State
-const updateState =  async (req, res, next) =>{
+const updateState = async (req, res, next) => {
     const _id = req.params.id;
-    const {name,  code, description, isActive } = req.body;
-    if( !name ||  !code ){
-        return res.status(400).json( { message: "Name and Code is required"});
+    const { name, code, description, isActive } = req.body;
+    if (!name || !code || !description) {
+        return res.status(400).json({ message: "Name, description and Code is required" });
     }
-    if( State.find( state => state.name === name || state.code === code  ) ){
-        return res.status(400).json( { message: "Name or Code is exists."});
-    }
+     
+    const existingState = await State.findOne({
+        $and: [
+          { name: { $regex: `^${name.trim()}$`, $options: 'i' } },
+          { _id: { $ne: _id } }
+        ]
+      });
+    if (existingState) {
+        return res.status(406).json({ message: "Name or Code already exists.", success: false });
+    } 
+    
     try {
-        const state  =  State.findByIdAndUpdate( { id: _id} , { name: name, code: code, description: description, isActive: isActive });
-        if(!state) {
-            return res.status(404).json( { message: `State with id : ${_id} not found.`});
+        const updatedState = await State.findByIdAndUpdate( _id , { name: name.trim(), code: code, description: description.trim(), isActive: isActive });
+        if (!updatedState) {
+            return res.status(404).json({ message: `State with id : ${_id} not found.` });
         }
-        await state.save();
+        
         return res.status(201).json({
             message: "State updated Successfully",
-            data: state,
+            data: updatedState,
             success: true
         });
 
@@ -120,22 +155,61 @@ const updateState =  async (req, res, next) =>{
     }
 
 }
+//to update a State status
+const updateStateStatus = async (req, res, next) => {    
+    const _id = req.params.id;  
+    const { isActive } = req.body;
+    
+    
+    if (isActive === undefined || isActive === null  ) {
+        return res.status(400).json({ message: "isActive is required" });
+    }
+    const existingState = await State.findById(_id);    
+    if (!existingState) {
+        return res.status(404).json({ message: `State with id : ${_id} not found.` });  
+    }   
+    // console.log("isActive: ", req.body); 
+    // console.log("existingState: ", existingState);
+    try {
+        const updatedState = await State.findByIdAndUpdate(_id , { isActive: isActive }, { new: true });
+        if (!updatedState) {
+            return res.status(404).json({ message: `State with id : ${_id} not found.` });
+        }
+        return res.status(201).json({
+            message: "State updated Successfully",
+            data: updatedState,
+            success: true
+        });
+
+    } catch (error) {
+        console.log(error);
+        next(error)
+    }   
+
+}
 
 //to delete a State
-const deleteState =  async (req, res, next) =>{
+const deleteState = async (req, res, next) => {
     const _id = req.params.id;
     try {
-        const delState = State.findByIdAndDelete( _id);
-        if(!deleteState) {
-            return res.status(404).json( { message: `State with id : ${_id} not found.`});
+
+        //check the state is already assigned to any location   
+        // const location = await Location.find(location => location.stateId === _id);
+        const location = await Location.findOne({ stateId: _id });
+        if (location) {
+            return res.status(400).json({ message: "State is already assigned to a location." });
         }
-        await delState.save();
+        const delState = await State.findByIdAndDelete(_id);
+        if (!deleteState) {
+            return res.status(404).json({ message: `State with id : ${_id} not found.` });
+        }
+        // await delState.save();
         return res.status(201).json({
             message: "State deleted Successfully",
             data: delState,
             success: true
         });
-        
+
     } catch (error) {
         next(error)
     }
@@ -147,5 +221,6 @@ module.exports = {
     getStateById,
     createState,
     updateState,
-    deleteState
+    deleteState,
+    updateStateStatus
 }
